Now we don't round off times to the nearest second

+++ src/rec/slow/StretchyPlayer.h
@@ -26,7 +26,7 @@ class StretchyPlayer : public Listener<const VirtualFile&>,
   // Callback when we're ready to actually jump to a new time.
   virtual void operator()(const float& time);
 
-  void setTime(int time) { timeLocker_->set(time); }
+  void setTime(float time) { timeLocker_->set(time); }
 
   Listener<const VirtualFile&>* fileListener() { return &fileListener_; }
   AudioTransportSourcePlayer* getTransport() { return transportSource_.get(); }
